fix: float16 overflow in training ACC; add warmup scheduler + per-var logging to wb2 smokes
- trainerERA5gen2: call metrics(y_pred.float(), y) during training so
  torch.sum(pred_prime**2) over ncol=777k SE nodes does not overflow
  float16 (max 65504) → denominator=Inf → ACC=0 for every batch
- smoke_wb2_*: enable linear-warmup-cosine scheduler (3-step warmup,
  10 total steps) and save_metric_vars=True to log per-variable losses

+++ credit/trainers/trainerERA5gen2.py
@@ -243,7 +243,7 @@ class TrainerERA5Gen2(BaseTrainer):
 
             # collect metrics
             if y_pred is not None and y is not None:
-                metrics_dict = metrics(y_pred, y)
+                metrics_dict = metrics(y_pred.float(), y)
                 for name, value in metrics_dict.items():
                     value = torch.Tensor([value]).to(self.device, non_blocking=True)
                     if self.distributed:
